extracted formatter logic from print.h
